tweaked relation dependent thought giver
